Fixes in the main scene and main menu

+++ Assets/Scripts/PlayerStats.cs
@@ -16,16 +16,16 @@ public class PlayerStats : MonoBehaviour
 
     [Header("Stats")] // the stats - Nino
     public float damage = 1; 
-    public int hp = 4;
-    public int speed = 5; 
-    public int dash_chargers = 1;
+    public float hp = 4;
+    public float speed = 5; 
+    public float dash_chargers = 1;
     public float dash_coldown_reduction = 1;
     public float attack_speed = 1;
 
     public bool upgraded = false;
 
 
-   public void Update() // Refrences funktions to update the stats in verius scripts - Nino
+    public void Update() // Refrences funktions to update the stats in verius scripts - Nino
     {
         if(upgraded == true)
         {
